GBNS-104: ProductRequests 네이밍 수정
